Small refactors to property table
