+++ x-pack/plugins/apm/public/components/app/TransactionDetails/Transaction/TransactionPropertiesTableForFlyout.tsx
@@ -4,49 +4,23 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { EuiSpacer, EuiTab, EuiTabs } from '@elastic/eui';
-import { Location } from 'history';
+import { EuiSpacer, EuiTitle } from '@elastic/eui';
 import React from 'react';
 import { Transaction } from '../../../../../typings/es_schemas/ui/Transaction';
-import { fromQuery, history, toQuery } from '../../../shared/Links/url_helpers';
 import { PropertiesTable } from '../../../shared/PropertiesTable';
 
 interface Props {
-  location: Location;
   transaction: Transaction;
 }
 
 export const TransactionPropertiesTableForFlyout: React.SFC<Props> = ({
-  location,
   transaction
 }) => {
-  const metadataTab = { key: 'metadata', label: 'Metadata' };
-  const tabs = [metadataTab];
-  const currentTab = metadataTab;
-
   return (
     <div>
-      <EuiTabs>
-        {tabs.map(({ key, label }) => {
-          return (
-            <EuiTab
-              onClick={() => {
-                history.replace({
-                  ...location,
-                  search: fromQuery({
-                    ...toQuery(location.search),
-                    flyoutDetailTab: key
-                  })
-                });
-              }}
-              isSelected={currentTab.key === key}
-              key={key}
-            >
-              {label}
-            </EuiTab>
-          );
-        })}
-      </EuiTabs>
+      <EuiTitle size="s">
+        <h4>Metadata</h4>
+      </EuiTitle>
       <EuiSpacer />
       <PropertiesTable item={transaction} />
     </div>
